Fixing the NGAP ASN.1 specs to address the issue reported with #12846
The issue reported with #12846 is caused by the missaligment of the
NGAP R16 ASN.1 file used to generate the NGAP message library with the
published standard. The file was previously modified to avoid an error
that the asn1 compiler was throwing for the missing external reference.

The fix is returning the NGAP R16 ASN.1 to the published standard and
is dereferencing the `ProtocolIE-SingleContainer` to avoid the compiler
error (untill the compiler issue is addressed).

All sanity test had passed.

Signed-off-by: Jordan Vrtanoski <jordan.vrtanoski@gmail.com>