feat(pipelined) Adding basic enforcement of the UE_AMBR
In the 3GPP standars it's defiend that the E-UTRAN might/could enforce the UE_AMBR , however this enforcement is not mandatory. This feature allowes the enforcement of the UE_AMBR which is derived from the subscriber data plan to be enforced on the EPC side without the need to create additional PCRF enforcement rules.

Signed-off-by: Jordan Vrtanoski <jordan.vrtanoski@gmail.com>